pyproject.toml
[project]
name = "cached-llm"
version = "0.0.7"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "diskcache>=5.6.3",
    "openai>=1.101.0",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.1.1",
]
